requirements.txt
# Core dependencies
Flask
torch
torchvision
Pillow
numpy
Werkzeug

# Optional: For Streamlit version
streamlit
